Rename `tests.rs` to `old_tests.rs` to avoid naming conflicts
