+++ zrml/prediction-markets/src/lib.rs
@@ -26,8 +26,8 @@ extern crate alloc;
 mod benchmarks;
 pub mod migrations;
 pub mod mock;
+mod old_tests;
 pub mod orml_asset_registry;
-mod tests;
 pub mod weights;
 
 pub use pallet::*;
